feat(DTFS2-7367): move e2e tests into cancellation disabled folder
