+++ e2e-tests/tfm/cypress/e2e/journeys/case/deal-cancellation-disabled/activities/activities.spec.js
@@ -1,11 +1,11 @@
-import relative from '../../../relativeURL';
-import { errorSummary } from '../../../partials';
-import activitiesPage from '../../../pages/activities/activitiesPage';
-import activityCommentBoxPage from '../../../pages/activities/activityCommentBoxPage';
-import MOCK_DEAL_AIN from '../../../../fixtures/deal-AIN';
-import { BUSINESS_SUPPORT_USER_1, BANK1_MAKER1, ADMIN } from '../../../../../../e2e-fixtures';
+import relative from '../../../../relativeURL';
+import { errorSummary } from '../../../../partials';
+import activitiesPage from '../../../../pages/activities/activitiesPage';
+import activityCommentBoxPage from '../../../../pages/activities/activityCommentBoxPage';
+import MOCK_DEAL_AIN from '../../../../../fixtures/deal-AIN';
+import { BUSINESS_SUPPORT_USER_1, BANK1_MAKER1, ADMIN } from '../../../../../../../e2e-fixtures';
 
-import { todayFormatted } from '../../../../../../e2e-fixtures/dateConstants';
+import { todayFormatted } from '../../../../../../../e2e-fixtures/dateConstants';
 
 context('Users can create and submit comments', () => {
   let dealId;
