+++ e2e-tests/tfm/cypress/e2e/journeys/case/deal-cancellation-disabled/tasks/tasks-AIN.spec.js
@@ -1,9 +1,9 @@
-import relative from '../../../relativeURL';
-import partials from '../../../partials';
-import pages from '../../../pages';
-import MOCK_DEAL_AIN from '../../../../fixtures/deal-AIN';
-import * as MOCK_USERS from '../../../../../../e2e-fixtures';
-import { BANK1_MAKER1, ADMIN } from '../../../../../../e2e-fixtures';
+import relative from '../../../../relativeURL';
+import partials from '../../../../partials';
+import pages from '../../../../pages';
+import MOCK_DEAL_AIN from '../../../../../fixtures/deal-AIN';
+import * as MOCK_USERS from '../../../../../../../e2e-fixtures';
+import { BANK1_MAKER1, ADMIN } from '../../../../../../../e2e-fixtures';
 
 const { T1_USER_1 } = MOCK_USERS;
